backend/python_model/requirements.txt
# backend/python_model/requirements.txt

flask
flask-cors
ultralytics
Pillow
numpy